// utils/vars/enums.ts
export enum PersonalInformationSteps {
	RETAIL_CLIENT = 1,
	INVESTMENT_CAT = 2,
	BIO = 3,
	CONTACT_INFO = 4,
	EMPLOYMENT_INFO = 5,
	BANK_INFO = 6,
    IDENTITY_PROOF = 7,
    RISK_PROFILE = 8,
	REVIEW = 9
}


export enum NextOfKinSteps
{
    BIO = 1,
    CONTACT = 2,
    PROOF_OF_IDENTITY = 3,
    REVIEW = 4
}


export enum DisclosuresSteps
{
    SIGNATURE_UPLOAD = 1,
    BLIND_ILLITERATE = 2,
    PEP = 3,
    FATCA = 4,
    KESTREL_TERMS = 5,
    KESTREL_NOMINEE_AGREEMENT = 6,
    AFRIVEST_EMAIL_INDEMNITY = 7,
    DECLARATIONS = 8,
    AFRIVEST_PRIVACY_POLICY = 9,
    REVIEW = 10
}